Exclude nodes from external load balancers in Kubernetes 1.19
Starting in Kubernetes 1.19, unschedulable nodes were included in load
balancers again (see kubernetes/kubernetes#90823), which causes a
conflict between the cloud controller and lifecycle-manager, where
lifecycle-manager attempts to remove the instance from the load
balancer, but then the cloud controller attempts to re-add it.

Kubernetes 1.19 also moved the ServiceNodeExclusion feature gate to beta
support, and defaults it to enabled, so you can set the
node.kubernetes.io/exclude-from-external-load-balancers label on the
node to have the node be excluded. An older version of this feature used
the alpha.service-controller.kubernetes.io/exclude-balancer, but this
label is removed in Kubernetes 1.19.

This commit sets the new
node.kubernetes.io/exclude-from-external-load-balancers label in
addition to the older alpha label so lifecycle-manager continues to work
on both old and new clusters.

Signed-off-by: Sarah Hodne <sarah@circleci.com>